Feature2.js -> read_data.js , Feature2.js
